requirements.txt
PyGithub==2.1.1
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4 